requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.22
botocore==1.31.22
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
decouple==0.0.7
Django==3.2
django-apscheduler==0.6.2
django-cors-headers==4.2.0
django-crontab==0.7.1
django-storages==1.13.2
djangorestframework==3.12.0
drf-yasg==1.21.5
frozenlist==1.3.3
google-auth==2.22.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.24.3
openai==0.27.8
packaging==23.1
pandas==2.0.1
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
s3transfer==0.6.1
scipy==1.11.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
tzlocal==5.0.1
uritemplate==4.1.1
urllib3==2.0.2
xmltodict==0.13.0
yarl==1.9.2
